+++ lib/src/script.dart
@@ -145,7 +145,7 @@ class Script {
   ///
   /// [the README]: https://github.com/google/dart_cli_script/blob/main/README.md#argument-parsing
   factory Script(String executableAndArgs,
-      {List<String>? args,
+      {Iterable<String>? args,
       String? name,
       String? workingDirectory,
       Map<String, String>? environment,